aplicacion de estilos al formulario

+++ src/components/components-alta/TablaFila.jsx
@@ -1,14 +1,32 @@
 import { useContext } from "react"
 import ProductosContext from "../../contexts/ProductosContext"
+import Swal from "sweetalert2";
+import './TablaFila.scss'
 
 const TablaFila = ({ producto }) => {
 
   const { eliminarProductoContext, setProductoAEditar } = useContext(ProductosContext)
 
   const handleEliminar = (id) => {
-    // Lógica de Sweet Alert
-    eliminarProductoContext(id)
-  }
+    Swal.fire({
+      title: "Estás eliminando un producto ¿Deseas continuar?",
+      text: "No podrás revertir esta acción",
+      icon: "warning",
+      showCancelButton: true,
+      confirmButtonColor: "rgb(146, 28, 28)",
+      cancelButtonColor: "rgb(146, 28, 28)",
+      confirmButtonText: "Eliminar",
+    }).then((result) => {
+      if (result.isConfirmed) {
+        eliminarProductoContext(id);
+        Swal.fire({
+          title: "Producto Eliminado",
+          text: "El producto fue eliminado",
+          icon: "success",
+        });
+      }
+    });
+  };
 
   const handleEditar = (producto) => {
     setProductoAEditar(producto)
@@ -17,7 +35,7 @@ const TablaFila = ({ producto }) => {
   return (
     <tr>
         <td>{producto.nombre}</td>
-        <td>{producto.precio}</td>
+        <td>${producto.precio}</td>
         <td>{producto.stock}</td>
         <td>{producto.categoria}</td>
         <td>{producto.descripcion}</td>
@@ -26,9 +44,9 @@ const TablaFila = ({ producto }) => {
         </td>
         <td>{producto.envio ? 'si' : 'no'}</td>
         <td>
-            <button>Ver</button>
-            <button onClick={() => handleEditar(producto)}>Editar</button>
-            <button onClick={() => handleEliminar(producto.id)}>Borrar</button>
+            <button className="boton-ver">Ver</button>
+            <button className="boton-editar" onClick={() => handleEditar(producto)}>Editar</button>
+            <button className="boton-borrar" onClick={() => handleEliminar(producto.id)}>Borrar</button>
         </td>
     </tr>
   )
